Add WorkNetNode class
